auth/login: move oauth related functionality
